+++ yash-semantics/src/command/compound_command/for_loop.rs
@@ -220,11 +220,7 @@ mod tests {
         let mut env = Env::new_virtual();
         env.builtins.insert(
             "check",
-            Builtin {
-                r#type: yash_env::builtin::Type::Mandatory,
-                execute,
-                is_declaration_utility: Some(false),
-            },
+            Builtin::new(yash_env::builtin::Type::Mandatory, execute),
         );
         let command: CompoundCommand = "for i in 1; do check; done".parse().unwrap();