+++ yash-semantics/src/trap/signal.rs
@@ -213,11 +213,7 @@ mod tests {
         let mut env = Env::with_system(Box::new(system.clone()));
         env.builtins.insert(
             "check",
-            Builtin {
-                r#type: yash_env::builtin::Type::Mandatory,
-                execute,
-                is_declaration_utility: Some(false),
-            },
+            Builtin::new(yash_env::builtin::Type::Mandatory, execute),
         );
         env.traps
             .set_action(